El Nino Oscillation
This is an assignment from my data science class. We had to plot the 6 strongest ENSO events from 1870-1950. This includes taking the website, making it into a text file, running it through pandas, sets the limit for each ENSO event to be 3 years and sets the start and stop time depending on numerical data, and plots the graph.